LPD-40224 Add a useNamespace attribute to the resourceURL tag so the URL doesn't get too big because of all the parameters
# breaking

## What util-taglib/src/META-INF/liferay-portlet-ext.tld

util-taglib/src/META-INF/liferay-portlet-ext.tld actionURL and resourceURL tags now include a new useNamespace attribute.

## Why

It was necessary to remove the portlet namespace from the parameters in a resourceURL tag. When there are a lot of parameters and they all get a namespace the URLs can exceed 2083 characters, and when this happens, it's possible to return a 400 Bad Request errors when using certain web servers and/or load balancers.

----

# breaking

## What util-taglib/src/com/liferay/taglib/portlet/ActionURLTag.java

util-taglib/src/com/liferay/taglib/portlet/ActionURLTag.java now includes a new _useNamespace attribute in the class and also in the doTag method parameters.

## Why

It was necessary to remove the portlet namespace from the parameters in a resourceURL tag. When there are a lot of parameters and they all get a namespace the URLs can exceed 2083 characters, and when this happens, it's possible to return a 400 Bad Request errors when using certain web servers and/or load balancers.

----